+++ packages/babel-plugin-i18n-calypso/package.json
@@ -20,5 +20,8 @@
   },
   "keywords": [],
   "author": "Alex Kirk <alex.kirk@automattic.com> (https://automattic.com)",
-  "license": "GPL-2.0+"
+  "license": "GPL-2.0+",
+  "scripts": {
+    "build": "node ../../bin/build-package"
+   }
 }
